// src/Interface/Post.interface.ts
// export interface PostInterface{
//     id: number;
//     img:string;
//     title: string;
//     desc: string
// }
// Generated by https://quicktype.io

export interface PostResponse {
    data: PostInterface;
}

export interface PostInterface {
    userId?:    string;
    title:     string;
    desc:      string;
    image:     string;
    publish?:   string;
    cat?:       string;
    avaliable?: boolean;
    _id:       string;
    createdAt?: string;
    updatedAt?: string;
    __v?:       number;
}